// src/utils/metadata/types/sitemap.ts
export type Sitemap = {
  url: string;
  priority?: number;
  images?: string[];
  videos?: Video[];
  alternates?: Alternates;
  lastModified?: string | Date;
  changeFrequency?: 'always' | 'hourly' | 'daily' | 'weekly' | 'monthly' | 'yearly' | 'never';
}[];

type BooleanString = 'yes' | 'no';
type Permission = 'allow' | 'deny';

export type Video = {
  tag?: string;
  title: string;
  rating?: number;
  duration?: number;
  view_count?: number;
  description: string;
  player_loc?: string;
  uploader?: Uploader;
  live?: BooleanString;
  content_loc?: string;
  thumbnail_loc: string;
  platform?: Restriction;
  restriction?: Restriction;
  family_friendly?: BooleanString;
  expiration_date?: Date | string;
  publication_date?: Date | string;
  requires_subscription?: BooleanString;
};

type Uploader = {
  info?: string;
  content?: string;
};

type Restriction = {
  content: string;
  relationship: Permission;
};

export type Alternates = {
  languages?: Languages<string>;
};

type Languages<T> = {
  [key in HrefLang]?: T;
};